ext: hal: nordic: Update nrfx to version 1.0.0
Updates nrfx to the recently released version and adjusts its glue
layer accordingly.
License headers are slightly corrected to match the license text at
https://opensource.org/licenses/BSD-3-Clause.
Deprecated MDK files nrf51xxx_peripherals.h are removed, since they
are replaced by a common file named nrf51_peripherals.h.

Origin: nrfx
License: BSD 3-Clause
URL: https://github.com/NordicSemiconductor/nrfx/tree/v1.0.0
commit: cf78ebfea1719d85cf4018fe6c08cc73fe5ec719
Purpose: Provide peripheral drivers for Nordic SoCs
Maintained-by: External


Former-commit-id: 3dd8b903d23e9e14df5badb15b2677d2c18a81e4

+++ ext/hal/nordic/nrfx/drivers/src/nrfx_adc.c
@@ -3,7 +3,7 @@ DECL|NRFX_LOG_MODULE|macro|NRFX_LOG_MODULE
 DECL|adc_cb_t|typedef|} adc_cb_t;
 DECL|adc_sample_process|function|static bool adc_sample_process()
 DECL|event_handler|member|nrfx_adc_event_handler_t event_handler;
-DECL|idx|member|uint8_t idx;
+DECL|idx|member|uint16_t idx;
 DECL|m_cb|variable|m_cb
 DECL|nrfx_adc_buffer_convert|function|nrfx_err_t nrfx_adc_buffer_convert(nrf_adc_value_t * buffer, uint16_t size)
 DECL|nrfx_adc_channel_disable|function|void nrfx_adc_channel_disable(nrfx_adc_channel_t * const p_channel)
@@ -17,5 +17,5 @@ DECL|nrfx_adc_uninit|function|void nrfx_adc_uninit(void)
 DECL|p_buffer|member|nrf_adc_value_t * p_buffer;
 DECL|p_current_conv|member|nrfx_adc_channel_t * p_current_conv;
 DECL|p_head|member|nrfx_adc_channel_t * p_head;
-DECL|size|member|uint8_t size;
+DECL|size|member|uint16_t size;
 DECL|state|member|nrfx_drv_state_t state;
